fix: resolve clippy warnings and link errors for vcpkg static ffmpeg
- Add build.rs to link Windows MF SDK libs (mfuuid, strmiids, mfplat)
  required by vcpkg static ffmpeg's avcodec Media Foundation encoder
- Add Eq derive to BooruFilter (required by SubscriptionState's Eq)
- Fix unnecessary cast and redundant closure clippy warnings
- Add booru_filter column to test DB schema

+++ src/scheduler/booru_engine.rs
@@ -133,7 +133,7 @@ impl BooruEngine {
 
         let posts = site_ctx
             .client
-            .get_posts(tags, site_ctx.config.page_limit as u32, 1)
+            .get_posts(tags, site_ctx.config.page_limit, 1)
             .await
             .context("Failed to fetch posts from booru")?;
